+++ src/utils.ts
@@ -1,4 +1,14 @@
-import { openAsBlob } from 'fs'
+import fs from 'fs'
+
+export const withResolvers = <T>() => {
+	let resolve: (value: T | PromiseLike<T>) => void
+	let reject: (reason?: unknown) => void
+	const promise = new Promise<T>((res, rej) => {
+		resolve = res
+		reject = rej
+	})
+	return { promise, resolve: resolve!, reject: reject! }
+}
 
 export const unwrapArrayIfSingle = <T extends unknown[]>(
 	x: T
@@ -23,39 +33,39 @@ export const readFileToWebStandardFile = (
 	for (let i = 0; i < files.length; i++)
 		buffers.push(
 			new Promise<File>((resolve, reject) => {
-				// if (openAsBlob)
-				resolve(
-					openAsBlob(files[i].filepath).then(
-						(blob) =>
-							new File([blob], files[i].originalFilename, {
-								type: files[i].mimetype,
-								lastModified:
-									files[i].lastModifiedDate.getTime()
-							})
+				if (fs.openAsBlob)
+					resolve(
+						fs.openAsBlob(files[i].filepath).then(
+							(blob) =>
+								new File([blob], files[i].originalFilename, {
+									type: files[i].mimetype,
+									lastModified:
+										files[i].lastModifiedDate.getTime()
+								})
+						)
 					)
-				)
-				// else {
-				// 	const buffer = Array<any>()
-				// 	const stream = createReadStream(files[i].filepath)
+				else {
+					const buffer = Array<any>()
+					const stream = fs.createReadStream(files[i].filepath)
 
-				// 	stream.on('data', (chunk) => buffer.push(chunk))
-				// 	stream.on('end', () =>
-				// 		resolve(
-				// 			new File(
-				// 				[new Blob([Buffer.concat(buffer)])],
-				// 				files[i].originalFilename,
-				// 				{
-				// 					type: files[i].mimetype,
-				// 					lastModified:
-				// 						files[i].lastModifiedDate.getTime()
-				// 				}
-				// 			)
-				// 		)
-				// 	)
-				// 	stream.on('error', (err) =>
-				// 		reject(`error converting stream - ${err}`)
-				// 	)
-				// }
+					stream.on('data', (chunk) => buffer.push(chunk))
+					stream.on('end', () =>
+						resolve(
+							new File(
+								[new Blob([Buffer.concat(buffer)])],
+								files[i].originalFilename,
+								{
+									type: files[i].mimetype,
+									lastModified:
+										files[i].lastModifiedDate.getTime()
+								}
+							)
+						)
+					)
+					stream.on('error', (err) =>
+						reject(`error converting stream - ${err}`)
+					)
+				}
 			})
 		)