+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.2.2 - 27 Dec 2024
+Bug fix:
+- Set minimum Elysia version to 1.2.7
+- Response doesn't reconcile when handler return `Response` is used with `mapResponse`
+- [#11](https://github.com/elysiajs/node/pull/11) `Content-Length` with `UTF-8` 2-byte characters
+- Ponyfill for `fs.openAsBlob` and `Promise.withResolve`
+
 # 1.2.1 - 25 Dec 2024
 Change:
 - Set minimum Elysia version to 1.2.3